Slow progress towards viable python.
- Jython is dead
- JEP requires python-side configuration
- GraalPython is a pain to use outside of graal.

I think I've converged on a sort of reasonable solution.  I grabbed the Python client code from Mimir Classic.  This spins up a python interpreter and invokes it on the __main__.py class in the resources directory.  Notably, this still works if __main__.py is in a jar file (jar = zip, and if you give python a zip file, it will call __main__.py at the root level of the zip).  This gets us a hook into the user's python interpreter.

From here, there's a simple Json-based protocol that lets us interact with the running python interpreter.  Currently, python overrides sys.stdout and sys.stderr with wrappers that package all sent messages with the appropriate Json surroundings (kind of like what happens in the pure python implementation), and then takes over stdin and stdout for use with this protocol.  stderr remains as a debugging channel (remember that what the client sees as stderr is actually packaging messages in Json and sending them over stdin).

This is a start, but by no means a complete implementation.  I still need to port over python-side code for the python client and python dataset client.  That should be relatively straightforward, but...

At any rate, my strategy here will be to use the STDIN/STDOUT channel to facilitate interactions with ExecutionContext.  This has the added benefit that it shouldn't be too hard to swap STDIN/STDOUT with a socket-based implementation or other process execution strategies, for example a chroot-jailed (or dockerized) python for hosted deployments.