Fixing a somewhat premature optimization
Only DONE (and FROZEN) cells have predictable effects on the scope.  As
soon as we hit a cell in a different state, the scope is no longer valid
and we can't use it to say anything about the subsequent cells' state.

(we will be able to do better for WAITING cells once we let modules
 declare optional provenance predictors)