WIP on spreadsheets
The way forward is making itself a little clearer.  There's a ton of
leftover cruft here.  Virtually all of the interesting logic now lives in
Executor (which I think I'll rename to something like "UpdateLayer").

This class maintains:
 - frame: A frame of reference
 - dag: A collection of UpdateRules, each defined over one or more ranges of one or more columns
 - subscriptions: A set of rows considered "interesting" to the user
 - data: A set of cells data values, either for an subscribed row, or for a direct dependency of a subscribed row.  data values are given as futures.
 - triggers: For each cell with a data value, this lists all of the cell's dependencies

The class is responsible for scheduling cell execution, which it does in a future.

A huge documentation pass is required here, as is a whole mess of testing.

Notably, UpdateDAG is not required anymore