WIP: Attacking ORM hell (re #182)
I think I've found most of the performance lags in GetWorkflow.  It looks
like we're down to ~0.1s (although it's on different hardware).

- Module.describe no longer does any database lookups.  Instead, the caller
  is expected to provide nearly all values that would normally get pulled
  out of the database.  For bulk lookups like Workflow.describe, these values
  can be retrieved for the entire database in a single lookup (and are then
  joined back with the workflow client-side.  This makes a handful of operations
  more complicated, but utility methods on cell can be used to provide nearly
  all of the arguments.
- Modules.describeAll no longer serves a meaningful purpose and has been
  merged back into Workflow.

One big potential performance win remains: Looking up table artifact schemas,
since currently ArtifactSummary.summarize requires a full artifact materialization
otherwise (the materialization itself isn't a pain, but the repeated SQL queries are).
I'm wondering whether it might not be a bad idea to just outright drop the
ArtifactSummary object.  We try to avoid making the SQLite file too big anyway by
putting large blobs in the artifact-associated file, so there's not really going
to be **that** much of a performance penalty at this point to just load the data field
when needed.

Related to this, it might be useful to cache schema information as part of the
Dataset object when it is first created.  We nearly always have the dataframe available
at this point (if not its contents).