Move heavy lifting out of the database block.
This branch starts an effort to move database heavy-lifting out of the transactional critical section.  The key change is that many of the Catalog describe() methods now return () => ___Description instead of ___Description.  The idea here is that, although the describe method may require database calls, we can move a lot of the expensive operations (e.g., calls into spark) to a continuation that is called after the transaction closes.