First steps towards a cleaner tentative edits (#re 196)
TentativeEdits is a mess.  It started off as a way to inject tentative
cells into the workflow, but it also accumulated a whole bunch of
state management to get around Rx's lack of support for a reactive fold.

The new version instead manages Modules in a linked list.  This is a
tiny bit more brittle, since we can't put a module into multiple
workflows, but I'm fairly certain that this wasn't possible to begin
with, so... yeah.  This change also means that Modules and
TentativeModules know where in the list they live (e.g., if they're the
last element in the list, updates use a different function), and have
the ability to more easily look up e.g., the preceding/succeeding real
module.

If it becomes a problem, we can factor the code out
into a wrapper class like we did before.  The result is still a
Reactive Buffer of Module -> WorkflowElement

The new WorkflowElement trait (implemented by Module, ArtifactInspector
and TentativeModule) provides most of the stuff that previously lived
in refreshBufferState()... except now it's all done cleanly over a
linked list.  It's still a bit hamfisted (e.g., we can propagate only
deltas through the visibleArtifact objects), but the overall result
results in fewer Rx hacks (like Vars of Rxs).

The code also now makes a much cleaner distinction between a module's
position in the original workflow and the display position.

Still TODO:
  - it looks like the code to detect replacements isn't working for non-head elements.