Also use single-file mode for local UNLOADs (closes #169)
When outputting line-oriented formats like JSON or CSV, Spark adopts the
HDFS convention of dumping each data partition into its own file.  This
is useful when using HDFS, but annoying for local applications (e.g.,
UNLOAD DATASET to ./Outputs/foo) (see #169)

We'd already added some code to support unloading into an artifact,
which by definition needs to be a single file.  This commit extends that
behavior to when the target file is local (i.e., when it is not a URL).