Reverting most of the stuff I've done towards #219
Supporting python UDFs requires pyspark.  It's dumb, but Spark tries to invoke pyspark on the system python to get its daemon (compared to Vizier's approach of embedding the libraries in the jar).

Instead of that, I'm going to fall back to using type annotations as hints.  If the type annotation is in a friendly, welcoming format, I'll have export_module map it to Spark's internal format and we can take it from there.