AnnotateWithSequenceNumber no longer barfs on literals.
It looks like SparkPartitionID() returns an Int, not a Long (as much of the AnnotateWithSequenceNumber code assumes).  This glitch is silently fixed when code is executed, but there are some code paths (e.g., inline evaluation over inline rows) where it isn't.  Cast the result of SparkPartitionID() to a long (for consistency) to fix it.