Started work to transition to Vega from VegaLite
This commit pushes a workable 70%-ish of the Vega schema definition in
a play-json compatible format.  I will note that this takes ~1500 lines
of scala compared to the autogenerated cruft needed to support Vega Lite
(see vizier-vega), which is big enough to literally make the scala
compiler barf about having too much code.

This commit also removes artifacts/Chart, which is the legacy pre-1.x
chart format, and appears to be completely unused anywhere in the code.