Reconnect button and fixes to tentative edits (closes #247, #179, #265)
Auto-reconnecting is impractical without a clean way to resolve state differences between the user's current state and the server's state.  That's out of scope for now.  Instead, this commit makes the DISCONNECTED status indicator more prominant and adds a reconnect menu item.

Prepending a new module to the front no longer crashes.

Spark (Hive, technically) expects the user.dir property to be an absolute path.  This is now set correctly.