Profiler implementation
Debrief:

In Artifact.scala, a property was added to the artifact using updateDatasetProperty.

- if "is_profiled" is not a property then it will be attached to the artifact.
- if "is_profiled" is a property then the profiler will run ONLY IF it is set to the boolean FALSE.

Once the profiler runs then the profiled information is then set to the value of the property "is_profiled". This will prevent the profiler from running multiple times on a single dataset.

*TEST CASES*:
Test cases are held in PublishArtifactSpec.scala. You must run the whole PublishArtifactSpec.scala otherwise the second profiler test case will fail. It will fail since the profiler has not been set as a property (which is done in the first profiler test case).