Prototype of ipyflow integration based on initial exploration (based on initial exploration by szimmermann and vdib)
The choice to vendor ipyflow rather than import it through wheel is based on its fairly extensive dependency chain.  Most of the dependencies (e.g., jupyter) are irrelevant to us.  Perhaps the right choice would be to pull it in through pip, but this way we can avoid conflicts and ensure that a valid version is always present.

Apart from that, this push does a bit of reorganization, moving core python functionality from info.vizier.commands.python to info.vizier.python