Safety: Drop implicit tag rendering
Although it *is* really nice to be able to write code without implicitly
calling .render on ScalaTag tags, this also means that the point where
the tag is rendered is (i) chosen semi-automatically, and (ii) not
immediately visible to someone reading the code.

This is dangerous because each call to .render creates an entirely new,
fully independent DOM node.  This is problematic because saving the tag
and saving the DOM node have very different semantics: DOM operations on a tag instantiate a new node to act on, while DOM operations on the
DOM node act on the node (as probably intended).  In particular, moving
to explicit rendering avoids the following bug:

```
object Foo
{
  val root = div("foo")
}
object Main
{
  document.body.appendChild(Foo.root)
  Foo.root.replaceInnerText("bar") // does not change the document.
}
```

This could be avoided with implicits with explicit typing:
```
  val root: dom.Node = div("foo")
```
... but defaulting to sensible mechanics is preferable.