Python show dataset now works on externally provided datasets containing a zero
Under some circumstances, JSON may decode floats as integers.  When this happens, python type assertions start complaining that the data value is unsafe for re-serialization when it gets echoed back to the server.

This fix generalizes python type assertions to allow ints to get passed in where a float would be expected.