Asynchronously created modules now inserted correctly (closes #319)
Fixed a dumb off-by-one error in TentativeEdits was inserting newly
created modules one position earlier than expected (when looking for
the preceding cell, it accidentally called .prev twice).

This issue was obscured by the fact that the typical code path for
module insertions is the one that replaces a TentativeModule (and
modules appended to the workflow were immune as well).

Ultimately, the issue affected any module that was inserted at a
non-tail position without replacing a Tentative module:
- Spreadsheets (created by clicking a button, not via TentativeModule)
- Modules created by another client of the same workflow

Specific changes include
- Removed a spurious .prev that would cause certain asynchronous module
  insertions to appear one cell above where they actually should have
  been until a reload.
- Better documentation for how TentativeEdits.onInsertOne works.
- Test cases for TentativeEdits
- Documenting why WorkflowElement.safePrev and .safeNext are 'safe'