Temporary workaround for #320
- Increase the spreadsheet socket outgoing message buffer size.  100k slots should be enough to keep up with most scrolling through the spreadsheet.

The *correct* way to fix this, is to change SpreadsheetExecutor/SingleRowExecutor to support notifying on full-row updates, rather than on a cell-by-cell basis (which it does right now).  This shouldn't be too hard, since SingleRowExecutor operates at the granularity of rows anyway... but I want to have the time to do it right (which is not the case right now OK).