Progress towards tearing out mimir
Moving to substrait-spark requires migrating to Spark 3.4, which breaks
an unpleasantly large number of things.  Progress is proceeding.

- Silencing a bunch of warnings around the use of the global execution
context.  This is a performance thing... I need to revisit it, but not
right now
- build_routes.sc now correctly labels headers
- Tearing out Mimir references, including references to Caveats and
annotations, as well as any relevant routes
- I'm going to keep the spark.caveats package, as there's a ton of glue
code in there to wrap around invoking spark and dumping its data out
into a presentable form.  At some point, however, we should pull this
out
- Removing the get annotation route requires disabling the caveat
modal... but going to accept that for now.