Progressing: Mimir torn out, code compiles but doesn't run
- Removed all references to caveats, and brought in some Mimir helper
functions that Vizier was relying on.  Everything compiles now.
- UNFORTUNATELY, substrait-scala depends on substrait-core (which wants
Jackson Databind 2.20), and Spark 3.4.2 (which wants Jackson Databind
2.14).  From the looks of it, Spark will not run with Databind > 2.15 😠.
In the long term, this will be less of a problem since Spark will live
in its own plugin, and we can do all sorts of fun classloadery things
there.  For now, it would be nice not to have to tear spark out
entirely.  Going to try to futz with the build later.